Set standard names

Many datasets have bad column names featuring accents, blank spaces, slash, etc. Set standard names set standard names, ie names which can be used in programming.

set_standard_names(.data, prefix = "var_")

Arguments

.data

a tibble

prefix

name of the prefix for names beginning with numbers

Value

a tibble with standardized column names

Examples

library(readxl)
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

read_excel(
path = system.file("extdata","panorama.xlsx", package = "tricky")
) %>%
set_standard_names()

#> # A tibble: 293 × 15
#>    ministere_porteur
#>                <chr>
#> 1                SPM
#> 2                SPM
#> 3                SPM
#> 4                SPM
#> 5                SPM
#> 6              MAEDI
#> 7              MAEDI
#> 8              MAEDI
#> 9               MCAS
#> 10              MCAS
#> # ... with 283 more rows, and 14 more variables: ministere_nom_complet <chr>,
#> #   nom_du_projet <chr>, projet_interministeriel <chr>,
#> #   description_et_objectifs_du_projet <chr>,
#> #   en_savoir_plus_sur_le_projet <chr>, debut <dbl>,
#> #   duree_previsionnelle_en_annee <chr>, phase_du_projet_en_cours <chr>,
#> #   cout_estime <chr>, cout_estime_par_tranche <chr>, zone_fonctionnelle <chr>,
#> #   financement_programmes <chr>,
#> #   lien_vers_les_projets_annuels_de_performances <chr>,
#> #   date_de_publication <dbl>